# openapi.yaml
openapi: 3.0.3
info:
  title: StackMonitor API
  description: |
    StackMonitor provides a comprehensive log monitoring and analytics platform.
    
    ## Features
    - **Real-time log ingestion** from multiple agents (Go, Python)
    - **Advanced filtering** by service, level, timestamp
    - **Statistical analysis** of log patterns
    - **Error rate metrics** with time-series data
    - **Live streaming** via WebSocket
    - **Natural language queries** for log analytics
    
    ## Architecture
    - **Agents**: Go and Python agents collect logs from various sources
    - **Ingestion Service**: Receives, deduplicates, and stores logs in ClickHouse
    - **API Server**: Provides REST and WebSocket endpoints for querying logs
    - **ClickHouse**: High-performance columnar database for log storage
    
    ## Authentication
    Currently, the API does not require authentication (POC mode).
    
  version: 1.0.0
  contact:
    name: StackMonitor Support
    email: support@stackmonitor.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: http://{host}:8080/api/v1
    description: Custom host
    variables:
      host:
        default: localhost
        description: API server hostname or IP address
...
paths:
...
  /logs/stream:
    get:
      tags:
        - Streaming
      summary: WebSocket stream for live logs
      description: |
        Establishes a WebSocket connection for real-time log streaming.
        The server continuously queries ClickHouse for new logs and pushes them to connected clients.
        
        **Protocol**: WebSocket (ws://)
        **Message Format**: JSON array of LogEntry objects
        **Interval**: Every 2 seconds
        
        **Connection Upgrade**:
        ```
        GET /api/v1/logs/stream HTTP/1.1
        Host: localhost:8080
        Upgrade: websocket
        Connection: Upgrade
        Sec-WebSocket-Key: x3JJHMbDL1EzLkh9GBhXDw==
        Sec-WebSocket-Version: 13
        ```
      operationId: streamLogs
      responses:
        '101':
          description: Switching Protocols - WebSocket connection established
          headers:
            Upgrade:
              schema:
                type: string
                example: websocket
            Connection:
              schema:
                type: string
                example: Upgrade
        '400':
          description: WebSocket upgrade failed
          content:
            text/plain:
              schema:
                type: string
                example: "WebSocket upgrade failed: invalid headers"
components:
...
  securitySchemes:
    # Placeholder for future authentication
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT-based authentication (not currently implemented).
        Future implementation will require a bearer token in the Authorization header.

# No security applied by default (POC mode)
# security: []
...